clients/python-flask/requirements.txt
flask
pyjwt[crypto]
requests